Refactor: TextButton(export default -> named export)
