+++ src/components/Button/TextButton.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import TextButton from './TextButton';
+import { TextButton } from './TextButton';
 
 const meta: Meta<typeof TextButton> = {
   title: 'components/TextButton',
