+++ src/components/Button/TextButton.tsx
@@ -2,8 +2,6 @@ import { ButtonHTMLAttributes, PropsWithChildren } from 'react';
 
 export type TextButtonProps = ButtonHTMLAttributes<HTMLButtonElement> & PropsWithChildren;
 
-const TextButton = ({ children, ...props }: TextButtonProps) => {
+export const TextButton = ({ children, ...props }: TextButtonProps) => {
   return <button {...props}>{children}</button>;
 };
-
-export default TextButton;
